refac: RepositoryFactoryパターンの再適用
