+++ src/api/hooks/card/useQueryCard.ts
@@ -1,17 +1,14 @@
-import { FetchError } from '@/api/clients/utils/fetchError'
-import { Card, cardFactory } from '../../models/card.model'
+import { FetchError } from '@/api/utils/fetchError'
+import { Card } from '../../models/card.model'
 import { useQuery } from '@tanstack/react-query'
+import { cardFactory } from '@/api/factory/card.factory'
 
 export const useQueryCards = (accessToken: string) => {
-  return useQuery<Card[], FetchError>(
-    ['cards'],
-    () => cardFactory().list(accessToken),
-    {
-      staleTime: Infinity,
-      onSuccess: () => {},
-      onError: (err: FetchError) => {
-        console.log(err)
-      },
-    }
-  )
+  return useQuery<Card[], FetchError>(['cards'], () => cardFactory().list(accessToken), {
+    staleTime: Infinity,
+    onSuccess: () => {},
+    onError: (err: FetchError) => {
+      console.log(err)
+    },
+  })
 }
